+++ rune/src/lint/rules.zig
@@ -101,6 +101,7 @@ const RULES = [_]RuleEntry{
     .{ .rule = .unique_prefix_redundancy, .handler = index_rules.checkUniquePrefixRedundancy },
     .{ .rule = .unique_index_redundant_with_fk, .handler = index_rules.checkUniqueIndexRedundantWithFk },
     .{ .rule = .unique_index_redundant_with_pk, .handler = index_rules.checkUniqueIndexRedundantWithPk },
+    .{ .rule = .unique_index_redundant_with_unique, .handler = index_rules.checkUniqueIndexRedundantWithUnique },
     .{ .rule = .table_no_index, .handler = index_rules.checkTableNoIndex },
     .{ .rule = .index_name_too_long, .handler = index_rules.checkIndexNameTooLong },
     // View validation rules (moved to view.zig)